Resolved Fixed/not Fixed Update issue.
Problem: If you do the raycast in Update(), it's happening more
often than it should.  If you do the raycast in FixedUpdate(),
then it's happening while Unity is in the middle of moving parts
to their new positions, and half the parts have moved and half
haven't yet, causing raycast to behave inconsistently.

Solution: Keep track using a flag as to whether or not a FixedUpdate()
has occurred after the previous Update().  Keep the work in Update(),
but have Update() abort and do nothing if no FixedUpdate() happened
yet after the last Update().  Thus only on the next Update() after a
FixedUpdate() will the mod do anything.